requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
Django==3.1.5
django-celery-beat==2.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.10
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.2.2
psycopg2-binary==2.8
django-redis==4.12.1
pytest-django==4.1.0
pylint==2.6.0